LRQA-43331 Move getFileContent into Workbench instead of PortalWorkbench
