+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/Workbench.java
@@ -29,6 +29,8 @@ public interface Workbench {
 
 	public File getDirectory();
 
+	public String getFileContent(String filePath);
+
 	public String getGitHubDevBranchName();
 
 	public String getGitHubURL();
